requirements.txt

streamlit
requests
pandas
numpy
python-dateutil
